requirements.txt
django==4.1.13
python-jose==3.3.0
python-dotenv==1.0.0
social-auth-app-django==5.4.0
pymongo==4.10.1
django-environ==0.11.2
requests==2.31.0
djangorestframework==3.15.1
mongoengine~=0.29.1
djongo~=1.3.6